+++ DistributionFactory/JobArrivalDistributionFactory.hpp
@@ -1,5 +1,5 @@
 //File:         JobArrivalDistributionFactory.hpp
-//Description:  A factory object that returns an implementation based on the selected distribution 
+//Description:  A factory object that returns an implementation based on the selected distribution
 //  *make_JobArrivalDistribution - returns a pointer to a Arrival Distribution
 //      choice:0 - Static Arrival
 //      choice:1 - Poisson Arrival
@@ -15,8 +15,8 @@ class JobArrivalDistributionFactory
 	public:
 	    static JobArrivalDistributionFactory *make_JobArrivalDistribution(DistributionType distribution);
 		virtual void Initialize() = 0;  //=0 because we have to declare some function;
-		virtual int getNext() = 0;      //=0 because we have to declare some function;
-	
+		virtual double getNext() = 0;      //=0 because we have to declare some function;
+
 };
 
 #endif
